Add an exception when creating reservation

+++ src/main/java/com/example/demo/controllers/ReservationController.java
@@ -1,5 +1,6 @@
 package com.example.demo.controllers;
 
+import com.example.demo.exceptions.reservation.ReservationNotAvailable;
 import com.example.demo.models.Box;
 import com.example.demo.models.DTO.ReservationDTO;
 import com.example.demo.models.Mapper.ReservationMapper;
@@ -10,6 +11,7 @@ import com.example.demo.services.BoxService;
 import com.example.demo.services.ReservationService;
 import com.example.demo.services.UserService;
 import java.util.ArrayList;
+import java.util.Objects;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,7 +21,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -62,7 +63,28 @@ public class ReservationController {
       if (user == null || box == null) {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
       }
-      Reservation reservation = Reservation.builder().id(new ReservationId(user, box)).user(user).box(box).reservation(reservationDTO.getReservation()).build();
+
+      // Check if there is enough space in the box
+      ArrayList<ReservationDTO> reservationsDTO = new ArrayList<>();
+      for (Reservation reservation : reservationService.getAll()) {
+        reservationsDTO.add(reservationMapper.toDTO(reservation));
+      }
+      int totalReservation = 0;
+      for (ReservationDTO reservation : reservationsDTO) {
+        if (Objects.equals(reservation.getBoxId(), reservationDTO.getBoxId())) {
+          totalReservation += reservation.getReservation();
+        }
+      }
+      if (totalReservation + reservationDTO.getReservation() > box.getQuantite()) {
+        throw new ReservationNotAvailable("");
+      }
+
+      Reservation reservation = Reservation.builder()
+          .id(new ReservationId(user, box))
+          .user(user)
+          .box(box)
+          .reservation(reservationDTO.getReservation())
+          .build();
       Reservation savedReservation = reservationService.create(reservation);
       ReservationDTO savedReservationDTO = reservationMapper.toDTO(savedReservation);
       return ResponseEntity.ok(savedReservationDTO);
@@ -72,22 +94,31 @@ public class ReservationController {
     }
   }
 
-
   @PutMapping(value = "/reservations/boxes/{boxId}/users/{userId}")
-  public ReservationDTO updateReservation(@PathVariable("boxId") Integer boxId,
-                                          @PathVariable("userId") Integer userId,
-                                          @RequestBody Reservation reservation) {
+  public ResponseEntity<ReservationDTO> updateReservation(@PathVariable("boxId") Integer boxId,
+                                                          @PathVariable("userId") Integer userId,
+                                                          @RequestBody Reservation reservation) {
     Box box = boxService.getById(boxId);
     User user = userService.getById(userId);
-    return reservationMapper.toDTO(reservationService.update(reservation,
-          new ReservationId(user, box)));
+
+    if (box == null || user == null) {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+    }
+
+    return ResponseEntity.ok(reservationMapper.toDTO(reservationService.update(reservation, new ReservationId(user, box))));
   }
 
   @DeleteMapping(value = "/reservations/boxes/{boxId}/users/{userId}")
-  public void deleteReservation(@PathVariable("boxId") Integer boxId,
-                                @PathVariable("userId") Integer userId) {
+  public ResponseEntity<Void> deleteReservation(@PathVariable("boxId") Integer boxId,
+                                                @PathVariable("userId") Integer userId) {
     Box box = boxService.getById(boxId);
     User user = userService.getById(userId);
+
+    if (box == null || user == null) {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+    }
+
     reservationService.delete(new ReservationId(user, box));
+    return ResponseEntity.noContent().build();
   }
 }
